requirements.txt
google-cloud-bigquery==3.10.0
pandas
db-dtypes
gcloud